[ADD] Added the possibility to use closure functions in the 'direct' mappings.
It allows to reduce the code in the @mappings methods as common modifiers can be
used. The change include 3 modifiers: one for converting the type of the source field,
one for getting the binding or openerp id of a backend field, one for getting the backend
id of an openerp id or binding id.